requirements.txt
appdirs==1.4.4
apturl==0.5.2
attrs==21.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
CacheControl==0.12.10
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cssselect==1.1.0
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
dell-recovery==0.0.0
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
et_xmlfile==2.0.0
fasteners==0.14.1
filelock==3.6.0
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
GDAL==3.4.1
geographiclib==2.0
geopy==2.4.1
greenlet==3.0.3
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipython_genutils==0.2.0
jeepney==0.7.1
Jinja2==3.0.3
jsonschema==3.2.0
jupyter-core==4.9.1
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.5.1
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.3
nbformat==5.1.3
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.5
OWSLib==0.25.0
packaging==21.3
pandas==2.2.2
paramiko==2.9.3
pbr==5.8.0
pexpect==4.8.0
Pillow==9.0.1
plotly==5.4.0
ply==3.11
protobuf==3.12.4
psycopg2==2.9.2
ptyprocess==0.7.0
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyproj==3.3.0
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.9.0.post0
python-debian===0.1.43ubuntu1
python-dotenv==1.0.1
pythran==0.10.0
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
QScintilla==2.11.6
reportlab==3.6.8
requests==2.25.1
scipy==1.8.0
scour==0.38.2
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==2.0.29
sympy==1.9
systemd-python==234
tenacity==6.3.1
tinycss2==1.1.1
tqdm==4.67.1
traitlets==5.1.1
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
webencodings==0.5.1
xdg==5
xkit==0.0.0
youtube-dl==2021.12.17
zipp==1.0.0
